requirements.txt
python-telegram-bot>=20.7
python-dotenv
yfinance
pandas
openai
feedparser
requests
dashscope